Simplify libfabric-oriented environment.
We had two environment variables for pointing at libfabric directories:
one specifying the include dir for the runtime build, and one specifying
the lib dir for linking user programs.  Here, switch to a single env var
for the libfabric install dir, because the include and lib paths are
both children of that and differ only in their last components.