Add an example of Chapel using Fortran functions and variables
This example defines some subroutines, functions, and variables in Fortran
and uses them in Chapel using 'extern' prototypes.

Since we don't have a header file with function prototypes in Fortran like
we do in C, we need to generate them ourselves.  Added a new pragma
"generate signature" that can be used on extern functions or variables and
tells codegen to generate prototypes for them even though they're extern.