Start on AM support.
The AM request type is filled out, at least for executeOn.  The AM
handler is outlined and running.  The common code for executeOn calls is
here, though everything is blocking for now.  On XC with 2 nodes and the
sockets provider we can send and receive AM requests and the AM handler
can send 'done' indicators, all without error, but we are getting an
error in the AM handler when we try to wait for the TxCQ completions
from sending 'done'.