Change strategy for comparing mandelbrot image output
Previously, we insisted on a binary match, but had some
prediffs that tried a few different images.

This commit changes it to instead run a PREDIFF script that
counts the number of bit mismatches and tolerates a small number
of these. It is set up to tolerate no mismatches for CLBG testing.