Check the CQ or the counter, as appropriate for this transmit context.
In the AM handler we were doing a PUT to send the AM 'done' indicator
back, but then we were checking for completion of that via the CQ.  But
the AM handler transmit contexts don't have completion queues associated
with them.  They have completion counters instead.  Here, differentiate
between transmit contexts with completion queues and counters, and do
the right thing for each.