Fix sub_test in python test directory
This sub_test was relying on a known relative path from it to util/cron/sub_test
but when the directory was moved, that relative path was incorrect.  Modify the
sub_test to rely on a structure that will not change by using $CHPL_HOME (note
that this assumed CHPL_HOME is set when start_test is called, which I think is a
fair assumption but it gives me some pause).