Update alias analysis tests to assume inferConstRefs runs
LICM alias analysis currently doesn't consider aliases from const-ref
arguments. That might be OK for that pass but isn't what we'll want to do
for an improved and generalized alias analysis.

One result of LICM not considering const-ref arguments is that if
inferConstRefs runs (and before this PR it normally would run only
when remoteValueForwarding ran - ie. not local runs) then LICM
would produce different aliasing results.

This commit adjusts the tests to either:
 * expect LICM not to report const-ref arg aliases
   (assuming inferConstRefs has run)
 * modify some ref argument values so inferConstRefs will not mark them
   const