Allocate full-length ddata at once
... when capturing a range-shaped iterable into an array.

When invoking `proc chpl__initCopy(ir: _iteratorRecord)` on a
range-shaped `_iteratorRecord`, that range is known since #10942.
Therefore we know how many elements we are going to get, so we can
allocate a _ddata with that many elements up front, without the need
to resize it as the elements come in.

While there:

* factor out the calls to __primitive("array_alloc"),

* remove a stray comment - OK-ed by DavidI.