Strengthen const checking
... by propagating FLAG_REF_TO_CONST upon a PRIM_MOVE
in a couple of cases where it wasn't propagated before:

* upon a 'move' from a PRIM_ADDR_OF of a constant or a ref-to-const,
* upon a 'move' from a ref-to-const.

While there, pass 'rhs' as an extra argument to a couple of functions
because it is already computed. Tidy up formatting.

Add the codes from #9982 as tests.

Fix const violations in existing code. The troublesome change in:

    test/library/packages/Collection/DequeParity.chpl

was made in #7322 for obscure reasons. Revert that here.

Testing: linux64 -futures, gasnet.

Resolves #9982.