Add a future to track an internal error
I discovered that sending a generic type as an argument to a function resulted
in an internal error instead of a user-facing one.  This sounded familiar, but
I wasn't finding an issue for it, so I opened one.

It's probably pretty easy to fix though there is the consideration that
sometimes I've triggered that error by implementing something poorly in the
compiler.