Change error when calling throwing task function to calling function within it
Prior to this change, if a non-throwing function called a throwing function
within a coforall, cobegin, or begin, we would complain about the call to the
task function and use its name instead of the name of the function being called
within the task function.  This changes allows us to point to the actual call.

Basically, detects if a non-throwing function calls a task function marked as
throws.  If so, track how deep into that function we go and once we reach the
first call to a throwing function that is not a task function, act as though
the task functions weren't present (but only in the case where we know we have
gone into it from a non-throwing function and need to indicate the error).