Make the error messages for int literals that are too big more consistent
The errors for overflows when converting a string to a uint(64) for an
integer literal don't have an AST around, so had to explicitly give the
filename and line number in the format string.  This led to them being after
the word "error:" instead of before it.  Fix this by creating an otherwise
unused  AST with the correct line and file information and passing it to
USR_ERROR.