Add two tests of syncs around throwing calls in non-throwing functions
- One covers the case where there isn't a try/try! around the sync, meaning the
error would get rethrown.  In relaxed mode, this should trigger and did not
until the changes that will get added in this PR.
- The other covers the case where the is one, to demonstrate that having these
structures around the sync do not trigger the error (this is probably a
duplicate of other tests, but I wanted to be certain)

The first did not work prior to this change, while the second ensures that the
change did not break good code.