Update TryStmt to track if it was inserted due to a sync block
Sync blocks rethrown the errors they have gathered using their inserted catch
statement, but our throws analysis assumed that try/catch statements would
either handle the error or have an additional throw statement that would be
checked - this meant that our analysis was not complaining about throw
statements or throws calls in non-throwing functions if the call or statement
happened to only be within a sync block.

This change alters TryStmt to track this fact about its creation via a private
field and a const method to access that field.  When the error handling pass
traverses the AST and performs these checks, it will now act as though the sync
try/catch statements are not present for the purpose of determining whether to
error about throwing errors in non-throwing functions.