Add compilerError overload for chpl_fix_thrown_error for other argument types
Prior to this change, we would get a resolution error on the call to this
function when trying to throw a function that wasn't a subtype of Error.  The
user could probably figure out this meant they were throwing a bad type, but it
is better to say so plainly and not using implementation specific functions.
This change adds an overload of chpl_fix_thrown_error which takes in types that
aren't subtypes of Error and will emit a nicer error message

I'd like to take another stab at outputing the type name in the message, but
wasn't sure how to get around the "unmanaged/borrowed" output.