Allow the declared return type to be generic
This compiler change makes computing the return type follow some of the
"unspecified return type" code paths in the case when the return type
is specified and is generic.

#### test/users/thom/returnGeneric.chpl

Once upon a time, in an SVN repository far, far away, during a group meeting
that we shall not name, we discussed the topic of "return an uninstantiated
generic class" and decided "to not support this for now".

That decision and the supporting arguments were documented in a .future by
the anonymous group leader in 3e8cc7d4bd. The .future was promptly overturned
by a non-leader into a non-future test to lock in the error message.

The now-extinct future **DID** mention that "we permit a similar thing
to be done for formal arguments and it's convenient" and that
"this could be a convenient feature" if only "time allowed down the road"
and "we could resolve semantic questions".

Both concerns are now addressed and the non-future is being overturned
once more from an error message to supported behavior.

While there, add .future tests for a couple of other issues.

Resolves #10184. Resolves #11330.