Enable forall loops over zippered loop expressions
If we have a loop expression in the leader position of a forall statement,
in the AST it is a call to chpl__loopexpr_iter(). It is converted by
convertIteratorForLoopexpr() to call directly the leader iterator these().

When the loop expression itself is zippered, chpl__loopexpr_iter() accepts
a tuple of all things within the zip() syntax. However, the leader iterator
these() accepts only the first one of those things.

This PR adds an adjustment to handle this situation, by extracting this first
thing from the tuple.

Long term, we need to design the AST to eliminate the need in
convertIteratorForLoopexpr() altogether.

De-futurize and beef up the corresponding test.

Resolves #10282.