Enable array task-private variables
Prior to this change, a task-private variable of an array type crashed
the compiler. This is because the compiler skipped some processing of
the taskInit/taskDeinit blocks of ShadowVarSymbols (and so some
PRIM_INIT calls were left in the AST unhandled). This is because
the ShadowVarSymbol type mixes the properties of VarSymbol and FnSymbol.
In particular, ShadowVarSymbol:initBlock()/deinitBlock() behave much like
any BlockStmt adjacent to the enclosing ForallStmt, while having their
parentSymbol pointing to the ShadowVarSymbol, not the enclosing FnSymbol.

This change ensures that the taskInit/taskDeinit blocks are included
in the traversals in replaceRuntimeTypePrims() and insertCasts().

While there, have insertCasts() avoid descending into nested symbols
other than ShadowVarSymbols, because it skips those nested ASTs anyway.

Add tests to exercise

Resolves #11422.