Improve environment variable handling in slurm-gasnet_ibv launcher (#11532)
This change passes all environment variables to the compute nodes when the gasnet_ibv launcher is being used, taking care to handle the character-locale variables such as LC_ALL properly. (The launcher is written in Perl, which sets those variable incorrectly when left to its own devices.)

While working on this launcher, it was noticed that some other launchers may have similar problems. Issue #11536 was created to track those.

Closes #11355