Futures about generating errors when passing remote records to extern procs
The following tests show that, currently, we don't balk at passing the
address of a remote record to an external C routine even though that
routine won't be able to do anything useful with the record since it
isn't sensical locally.  They request that when compiling with
--checks (the default), the compiler-generated logic that narrows the
pointer before sending it off to the remote routine does a locality
check and generates an error message if it isn't already local.  See
the .future files (all identical) for details.

The .good messages here propose an error message, but I'm open to
other ones as well (and the ones I've chosen are questionable when the
remote expression is anything more than a plain variable).  Having any
error message would be preferable to the current behavior in which a
bad pointer is likely to be dereferenced.