Implement initial AM support.
This gets basic AMs working in comm=ofi.  At least, with the sockets
provider all the hellos pass.  (With the gni provider everything seems
to hang; I'll tackle that next.)

The AM descriptor declarations are in the public chpl_comm_bundleData_t
in the implementation-specific chpl-comm-task-decls.h file now, which
lets us use the stack-based executeOn bundles directly on the initiating
side rather than making copies of them.  There is functional support for
blocking and non-blocking normal executeOn.  "Fast" executeOn is present
nominally but in reality is done the "slow" way: rather than calling the
executeOn-body function directly from the AM handler we initiate a task
to run it.  There is also support here for AM-mediated GET and PUT,
which are used in situations where RMA is not possible due to lack of
registration, and there are new calls to this support in the appropriate
situations.

Relatedly, for any node:address combination you can now get either the
owning node's memory registration descriptor or the remote key.  It used
to be that you could only get the descriptor on the owning node itself
or the remote key on some other node, but it turns out that it's nice to
be able to get either of these on either kind of node, for sanity
checking if nothing else.

While here, I reverted to manual progress now that the AM handler is
running, made some more improvements to debugging including adding the
ability to abort (with core dump) on error, and converted some assert()s
which were inherently debug-only into always-present CHK_TRUE()s.