Add a variable assignment to satisfy the compiler.
In a non-debug build I got "'mrDesc' may be used uninitialized" in the
fi_write() call in ofi_put() and the fi_read() call in ofi_get() because
the compiler could not tell that the program would halt before reaching
those calls if the logic flow path which didn't define mrDesc was the
one that happened to be taken.  Here, add a functionally unnecessary
assignment to mrDesc to satisfy the compiler.