comm=ofi: AMs now work with the gni provider.
We had AMs working with the sockets provider, but with the gni provider
they were hanging.  I had guessed this was due to a progress problem but
it turned out to be something else entirely.  We "broadcast" the global
variables during initialization by having the non-0 nodes GET them from
node 0.  With the gni provider at least some of these GETs were done by
AMs to node 0 to PUT the values back to bounce buffers on the original
nodes, because the variables there were not in registered memory.  These
AM PUTs on node 0 were done in tasks, rather than directly in the AM
handler, because that provided easier completion management using the
worker thread transmit context CQs rather than the simple completion
counter used for the AM handler's transmit context.  Alas, at least one
of these tasks ended up queued to the same Qthreads worker as the main
program task, also on node 0, which was sitting in the post-broadcast
PMI barrier waiting for the other nodes because "broadcasting" globals
involved no explicit work for it.

In fewer words, this was basically, though indirectly, caused by a task
monopolizing a Qthreads worker thread by spinning in the PMI barrier
without yielding.

Here, resolve this problem by implementing a PUT-based barrier based on
the one in comm=ugni.  Use the out-of-band PMI barrier during the early
comm layer startup, and then switch to the PUT-based one as soon as RMA
and AMs are functional.

Except one more wrinkle: use the out-of-band barrier even after RMA and
AMs are functional, if the barrier call comes from the main thread
(whose identity we record in the chpl_comm_init() call).  We do this to
prevent the main thread from allocating a transmit context, since those
are intended for use by the worker threads.  This doesn't interfere with
the solution above, because the global variable broadcast comes after we
start running the Chapel main program as a task, on a worker thread.