Remove the complex test, as complex literals are virtually non-existant
From what I can tell, the compiler does not use Immediates in the case of
complex numbers.  I think this is because we can't tell the difference between
"2 + 1i" when "i" means imaginary and "2 + 1i" when "i" is the name of a
symbol available to the current scope.  This means that we can't support default
values for complex arguments until we support CallExprs.