Fix so the default values check doesn't impact renamed exported functions
The param cname formal was removed but we still continued operating on it, which
led to my change encountering it even though it wasn't in the tree any more.
I don't think it makes sense to continue formal operations on it once it has
been removed, so insert a continue.

Also, change my warning about default values and C compilation to only trigger
when compiling using `--library`