build_configs: Cray-XC aarch64 should leave cray-mpich module alone
This morning the Cray-XC aarch64 module build broke because command
"module unload cray-mpich" failed.

The cray-mpich module caused a lot of problems when I was first trying
these builds on early versions of the Cray-XC aarch64 platform.
That's why that "module unload cray-mpich" was put in the build script
in the first place.

It does not seem to be needed any more.