Use memmove instead of memcpy for chpl_gen_comm_{get,put}
We short circuit comm if the remote node is the same as the current locale, but
we were using a memcpy to do the "put"/"get". However, puts/gets can operate on
overlapping data, so we really need to be using memmove. Note that the
individual comm layer get/puts routines were already using memmove for
same-node check, it was just the higher level wrapper that wasn't.

This fixes the record-self-assignment.chpl test.