Add on statement to atomic deinit() and fix spelling error
It turns out that the atomic_destroy() calls are no-ops for most
configurations we care about (e.g., intrinsics), so the lack of an
on-clause previously was not an issue.  Meanwhile, for
CHPL_ATOMICS=locks, we've never really done much in the way of
multi-locale testing.  Adding the on-clause here to make the code not
complain for this case.  Elliot and I are hypothesizing that it won't
add much overhead due to our general fast-path local-on optimizations.

Fix a spelling error that Elliot caught as well.