Use a seed to guarantee behavior and avoid failure mode
This test uses random numbers to compute the value of pi to one decimal place as
a way to exercise MPI_Reduce.  Usually the random numbers generate an average in
the appropriate range but sometimes they produced a number that rounded up to a
3.2 instead of to 3.1.  This commit adds a seed to the random numbers used,
which allows us to continue to test the functionality we care about without the
variability that could result in failure.