Update BigInteger docs to say unsigned literal overflow is a compiler error
Integer literals that are too large to fit in uint(64) are now compiler
errors instead of truncating the number to max(uint(64)).  Update the
BigInteger documentation to reflect this.

Also add a test showing that literals larger than max(int(64)) do not
overflow when cast to `bigint` as long as they are <= max(uint(64)).