Change some aspects of _toLeader
This PR is in support of my WIP on reduce expressions and ForallStmt.

* Move _toLeader-related processing from cleanupLeaderFollowerIteratorCalls(),
which happens towards the end of lowerIterators(), into the newly-created
cleanupLeaderIteratorCalls() close to the start of lowerIterators().
That way, lowerForallStmtsInline() is free to inline _toLeader calls,
which I need in my WIP. (This PR does not introduce such inlining.)
Without this change, the processing in cleanupLeaderFollowerIteratorCalls()
would no longer fire if such inlining happened. Details below.

* Remove `proc _toLeader(iterator: _iteratorClass)` .
It was invoked from `_toLeader(_iteratorRecord)`.
Have the latter do the work of the former directly, instead.
This is to simplify things slightly. It required the change in preFold.cpp.

An observant reader might ask...

* Q: How about doing the work that's now in cleanupLeaderIteratorCalls()
directly in preFold.cpp / the PRIM_TO_LEADER case?

  - A: That should be possible. The downside would be the introduction of
extra temps (see `ftemp`) at resolution, rather than later during
lowerIterators(). Also, this would not address the main issue with
the current _toLeader, which is referencing formals outside their
function. (See the comment there in preFold.cpp.)

* Q: How about making the same change for _toFollower and _toStandalone?

  - A: That should also be possible. I have not seen the need for that
yet in my WIP. It is likely that I will need it for _toStandalone
and not for _toFollower. Ideally we will elimnate the _to* machinery
altogether.