Simplify implementation of resolveInitVar
This commit refactors 'resolveInitVar' to clarify the various special
cases. The case when coercion is required is more easily identified by
checking for specific types rather than the absence of those types, as
well as using a boolean instead of whether a pointer is NULL.

The case where chpl__initCopy is required is clearer by checking for
specific types rather than using initCopy as a default.

This commit does change the AST, such that raw 'init' calls are now used
more for copy-initialization in favor of 'chpl__initCopy'. As a result,
some other minor fixes were required:
- update 'handleSetMemberTypeMismatch' to look at non-ref types
- using the type constructor as an alternative instantiation point in
  'findCopyInit'