Bump the timeout for parallel/sync/waynew/class3
For the last week, nightly gasnet-everything testing has been unable to
successfully kill this test, which has led to hangs in testing. This
future currently deadlocks, so it used a 3 second timeout to prevent
execution from using much time. I have been unable to reproduce this on
my own (even running as the nightly tester in the same workspace trying
to reproduce as closely as I can.) As a shot in the dark, try increasing
the timeout to maybe avoid a potential race in timedExec.