Revert "Don't start up the polling thread when only using gasnet with 1 node"
This reverts commit 2128c2fae6297f93393fdd8fa8e4cb3f11a11ad5.

For reasons that we don't understand, not starting the polling thread is
preventing us from killing a deadlocked parallel/sync/waynew/class3
under nightly gasnet udp testing. However, this hang is only
reproducible under jenkins and we have not been able to reproduce
outside of jenkins. For now, just revert this change, since it's not an
important change.