Fix getTheIteratorFn() on tuples
* When getTheIteratorFn() is invoked on a tuple of iterator classes,
return the iterator function corresponding to the first of them.
Previously it was returning the first IC's getIterator function,
which is not useful.

This required an adjustment in expandForLoop() (towards the lowerIterators.cpp)
to preserve the behavior of skipping expandIteratorInline() in the tuple case.

We see only one level of tuple-ness when combining multiple iterator classes
over our entire test suite. This is probably because syntactically a loop
can have only one level of zipper-ness in its iterable expression. If we
encounter multiple levels, getTheIteratorFn(Type*) can be made recursive.

* Remove getTheIteratorFnFromIteratorRec(). Instead, have getTheIteratorFn()
handle the iterator record case. An iterator record and an iterator class
point to the same IteratorInfo, so it is natural to handle them both along
the same code path.