Initial comm=ofi network atomic interface implementation.
This implements the comm layer network atomic interface for comm=ofi.
For now the operations are done using Active Messages and CPU atomic
ops.  Native network atomic ops will be added next, but those will need
AM-mediated CPU AMOs as a fallback for the cases they cannot cover, so
the latter had to be here first.

This passes test/runtime/configMatters/comm/*atomic*.chpl on a Cray XC
with both the sockets and gni providers, with certain limitations.  The
most notable of these limitations is that CHPL_RT_NUM_THEADS_PER_LOCALE
has to be set low enough that we don't try to create more transmit
contexts per scalable endpoint than the provider's limit, because
transmit contexts are currently private to threads and the default
thread count (the number of cores per node) is always larger than the
sockets provider's limit of 16 transmit contexts per scalable endpoint.
(This limitation applies to all comm layer activities, not just AMOs,
and will be relaxed in the future by allowing multiple threads to share
a transmit context.)

While here I also fixed a warning erroneously emitted when hugepages are
not used in single-node programs, fixed a few bugs in the AM handling
that was already here, and made some minor improvements to the debugging
support.