Allow users to specify node access for slurm launchers
This allows users to specify a node access for slurm launchers by
setting `CHPL_LAUNCHER_NODE_ACCESS`. Previously, we always forced
"exclusive" access, but this isn't always appropriate. Now, we default
to exclusive, but users can override by setting:

 - `CHPL_LAUNCHER_NODE_ACCESS=shared` to get `--share` access
 - `CHPL_LAUNCHER_NODE_ACCESS=unset` to get the system default access (i.e we
    won't throw anything). This is generally useful, and is needed for cases
    like NERSC where throwing --exclusive to salloc and srun will cause hangs.
    For scalability collection on NESRC we couldn't use slurm-srun before
    because of this limitation, which required us to have a lot more
    infrastructure to replicate launcher capabilities. We should now be able to
    just run with `CHPL_LAUNCHER_NODE_ACCESS=unset` and get rid of all the
    extra code.
 - `CHPL_LAUNCHER_NODE_ACCESS=exclusive` to get `--exclusive` access (default)

Slurm 17 changed "--share" to "--oversubscribe", but `--share` will continue to
work, so use that for maximum portability.

In addition to "shared" for shared access we also support "share",
"oversubscribe" and "oversubscribed" since "share"/"oversubscribe" are
the real slurm terms and people may be more used to those. However, they
are not documented and are just for convenience.