Make some usability improvements to slurm-gasnetrun_ibv launcher
- Don't require setting --walltime/CHPL_LAUNCHER_WALLTIME for interactive runs
- Throw --quiet for interactive runs to hide useless "salloc granted" messages
- Stop trying to remove never created sysFilename for batch submissions
  - quiet failed `rm` messages

This also updates docs to stop saying that CHPL_LAUNCHER_WALLTIME is
required for slurm, and recommends using CHPL_LAUNCHER_PARTITION instead
of the underlying slurm env vars. It also removes a "in the future note"
about slurm-srun being a good option for ibv -- we can update the docs
if/when that becomes true.