Changes reflecting review feedback.
In the runtime, remove a superfluous comment about adding future network
atomic support for other operations and types.

In the SSCA2_kernels.chpl test, change decisions involving whether or
not we have network atomics support which were based on 'ugni' (the only
setting that supported them before) to be based on 'none' (the only one
that doesn't support them now).