Fix a memory leak when A.pop_front() and friends make an array empty
The `dsiDestroyArr` function was checking if the domain's `numIndices` was
greater than 0 to decide if there were any elements that needed to be
destroyed.

When `numIndices` is 0 due to pop_front() or similar functions, there are
possibly still allocated elements since the `dataAllocRange` may still be
non-empty.  Check both `numIndices` and `dataAllocRange` to decide if there
are any array elements to destroy.