Improve warnings for unknown Fortran type and kind
Added the symbol we are looking up the Fortran type/kind for as an argument
to the lookup function so it can be used for file/line information in
warnings.

Added a set of symbols that have been warned about so each symbol only gets
each warning once.

Updated a test with the new warnings.