Stop requiring subprocess32 for our testing infrastructure
Previously, we brought in subprocess32, which is a backport of python
3's subprocess to python 2. We did this to get features out of 3 that we
wanted, but weren't available in python 2. subprocess is not pure python
and relies on C code, which has caused us several problems over time. We
have to be sure to build our cray module with all versions of python
that we want to support and recently we've started running into build
issues on CLE6 UP07.

The only thing we've actually used subproccess32 for is to bring
subprocess.check_output support to python 2.6, so instead of dealing
with the hassle just stop depending on subprocess32 and add a
check_output like wrapper. It's based on the run_command script in
util/chplenv and the 2.7 implementation of check_output.