Remove dtSymbol
We used to have 'symbol' type, available in the user space of Chapel programs.
It was unused and unusable. So removing it now.

dtSymbol was added in a3f94884 aka r5388 to implement closures (!!), noting
that it "can be used to represent the set of functions with a particular name."