Fix timeout detection for qsub systems
When we switched to using `&> $output_file` redirection instead of the
qsub `-o $output_file` option we stopped being able to detect timeouts
on qsub systems. This is because the message that we key off of for
timeouts was still being sent to the non-redirected qsub e/o files. This
updates chpl_launchcmd to send any non-redirected output to a .more file
with `-j oe -o $output_file.more`. We delete that file if execution was
successful with the `-R oe` option. If that file still exists after we
read our normal output, we will concatenate it to our testing output.
This allows us to properly detect timeouts again. It also prevents the
temporary qsub e/o files from cluttering the workspace, since they're
now deleted for successful runs.