Don't use the `-R oe` option for qsub
I was using this to delete the non-redirected output file if testing was
successful, but it's a newer PBSPro option, and isn't available on older
versions or on Moab/Torque.

Turns out we don't actually need to use it anyways. We're setting `-o`
to our tmpdir and it will be deleted so we won't leave empty qsub e/o
files around, and the file will be empty for successful runs so we don't
have to worry about it adding testing output for successful runs.