Init some internal-module global variables early
When a global param in an internal module has a simple initializer
or a global type alias is for a simple type, allow them to be used
even before their declarations would be resolved according to the
standard resolution order.

"Obvious" cases are, for example:

  param localeModelHasSublocales = false;
  type chpl_table_index_type = int;

This change eliminates the need to tune the placement of global params
within the internal modules, which we had to do in the past.
This was motivated by my upcoming work, where localeModelHasSublocales
needs to be referenced before its module gets resolved.

We could extend this feature to vars and consts, for example:

  var chpl__testParOn = false;

In addition to setting such a variable's type, we would also
need to change the placement of the corresponding "move" statement
so that it is executed before any regular module initialization
code is executed. This is to eliminate the possibility of access
before the variable is initialized.

We may want to extend these changes to user code.
The downside would be the possibility of surprising behavior.
For example, the compiler would suddently reject the program
when the type of a global variable is changed from int to bigint.