Find leader and standalone iterators earlier
So far we have relied on resolving _toLeader and _toStandalone
to find the parallel counterpart to a serial iterator.

With this change, the compiler seeks leader and follower immediately
after resolving the serial iterator. We record the results
in an IteratorGroup, which is associated, via a hashtable,
with the serial iterator, as well as with any of its parallel
counterparts that have been found.

The bodies of the leader/standalone are not resolved at that time.
This is because they may contain resolution errors. Those errors
should not be reported yet because they (the iterators) may end up
being unused. If the leader and/or the standalone is not found, it is
not an error at that time either, again because they may not be needed.

It would be nice if we can store this information in the IteratorInfo.
However, when the resolving takes place, IteratorInfos have not been
set up yet.

The correctness of the IteratorGroups is verified when those iterators
are used in forall loop statements.

Requires #11675.