Extend homogeneous tuple support to permit uint*type signatures
Prior to this, uint tuple types were not supported and small-uint
tuple types resulted in internal errors.

Here, the fix is to (a) add an overload for uint*type which does
a safe param downcast to int in module code and calls the normal
int*type overload; and (b) switch to using to_int() rather than
int_value() in tuples.cpp in order to support coercions from
types other than integers (note that this opens up support for
bool*type, for better or worse.  That loophole could presumably
be closed by adding another module overload, but I didn't bother
with that here (but could explore it if desired)).