Remove reference to homog. tuples being implemented via param int * type
The spec already had English text indicating that a homog. tuple could
be specified via a param integral * a type, but then specified a
specific signature that would be used to implement it.  With the
work in this PR, this is now inaccurate, and also feels like too much
information (more about how things are implemented, less about how
they should be).