Find leader and standalone iterators earlier (#11676)
Find leader and standalone iterators earlier

So far we have relied on resolving _toLeader and _toStandalone
to find the parallel counterpart to a serial iterator.

With this change, the compiler seeks leader and follower immediately
after resolving the serial iterator. We record the results in an
IteratorGroup. An IteratorGroup is stored in FnSymbol::iteratorGroup
of each iterator of the group. This is similar to FnSymbol::iteratorInfo.

The bodies of the leader/standalone are not resolved at that time.
This is because they may contain resolution errors. Those errors
should not be reported yet because they (the iterators) may end up
being unused. If the leader and/or the standalone is not found, it is
not an error at that time either - again because they may not be needed.

It would be nice to store pointers to leader and standalone
FnSymbols directly in the serial iterator's IteratorInfo,
instead of creating IteratorGroups.
This does not work :( because iterator-forwarding functions
do not have the associated IteratorInfo. Creating IteratorInfo
for iterator forwarders would require non-trivial changes to the compiler
because we often rely on fn->iteratorInfo != NULL to be an indicator
of fn being an iterator.

This PR does not change the compiler to rely on this information.
However, it adds verification that the standalone or leader are found
correctly during resolution of forall loop statements.

While there:  refactor the cleanup of IteratorInfo upon ~FnSymbol
and remove the unused IteratorTag and IteratorInfo::tag.

Requires #11675.

r: @mppf